tentando arrumar
